+++ config.go
@@ -4,12 +4,14 @@ import (
 	"github.com/BurntSushi/toml"
 	"os"
 	"path/filepath"
+	"strconv"
 )
 
 type Config struct {
-	Token  string
-	Domain string
-	User   int
+	Token   string
+	Domain  string
+	User    int
+	Aliases map[string]int
 }
 
 func (c *Config) Load(path string) error {
@@ -22,4 +24,12 @@ func (c *Config) Load(path string) error {
 	}
 
 	return nil
+}
+
+func (c *Config) Alias(v string) (int, error) {
+	if c.Aliases[v] != 0 {
+		return c.Aliases[v], nil
+	} else {
+		return strconv.Atoi(v)
+	}
 }