+++ CHANGELOG.md
@@ -2,6 +2,9 @@ WIP
 --
 
 * [CHANGE] DRYed up the code a bit
+* [NEW] Added support for environment aliases and a basic "export" aliases building from all repositories and
+  environments
+* [CHANGE] Use pointer instead of values all around to avoid shadowing too much
 
 v0
 --